+++ core-ng/src/main/java/core/framework/impl/log/stat/CollectStatTask.java
@@ -1,6 +1,6 @@
 package core.framework.impl.log.stat;
 
-import core.framework.impl.log.LogForwarder;
+import core.framework.impl.log.KafkaAppender;
 
 import java.util.Map;
 
@@ -8,17 +8,17 @@ import java.util.Map;
  * @author neo
  */
 public final class CollectStatTask implements Runnable {
-    private final LogForwarder logForwarder;
+    private final KafkaAppender kafkaAppender;
     private final Stat stat;
 
-    public CollectStatTask(LogForwarder logForwarder, Stat stat) {
-        this.logForwarder = logForwarder;
+    public CollectStatTask(KafkaAppender kafkaAppender, Stat stat) {
+        this.kafkaAppender = kafkaAppender;
         this.stat = stat;
     }
 
     @Override
     public void run() {
         Map<String, Double> stats = stat.collect();
-        logForwarder.forwardStats(stats);
+        kafkaAppender.forwardStats(stats);
     }
 }