+++ core-ng/src/main/java/core/framework/module/LogConfig.java
@@ -1,67 +1,38 @@
 package core.framework.module;
 
-import core.framework.impl.log.ActionLogger;
-import core.framework.impl.log.LogForwarder;
-import core.framework.impl.log.TraceLogger;
+import core.framework.impl.log.ConsoleAppender;
+import core.framework.impl.log.KafkaAppender;
 import core.framework.impl.log.stat.CollectStatTask;
 import core.framework.impl.module.ModuleContext;
 import core.framework.log.MessageFilter;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.nio.file.Path;
 import java.time.Duration;
 
 /**
  * @author neo
  */
 public final class LogConfig {
-    private final Logger logger = LoggerFactory.getLogger(LogConfig.class);
     private final ModuleContext context;
 
     LogConfig(ModuleContext context) {
         this.context = context;
     }
 
-    public void writeActionLogToConsole() {
-        if (context.isTest()) {
-            logger.info("disable action log during test");
-        } else {
-            context.logManager.actionLogger = ActionLogger.console();
+    public void writeToConsole() {
+        if (!context.isTest()) {
+            context.logManager.consoleAppender = new ConsoleAppender();
         }
     }
 
-    public void writeActionLogToFile(Path actionLogPath) {
-        if (context.isTest()) {
-            logger.info("disable action log during test");
-        } else {
-            context.logManager.actionLogger = ActionLogger.file(actionLogPath);
-        }
-    }
+    public void writeToKafka(String kafkaURI) {
+        if (!context.isTest()) {
+            KafkaAppender appender = KafkaAppender.create(kafkaURI, context.logManager.appName);
+            context.logManager.kafkaAppender = appender;
+            context.startupHook.add(appender::start);
+            context.shutdownHook.add(appender::stop);
 
-    public void writeTraceLogToConsole() {
-        if (context.isTest()) {
-            logger.info("disable trace log during test");
-        } else {
-            context.logManager.traceLogger = TraceLogger.console();
-        }
-    }
-
-    public void writeTraceLogToFile(Path traceLogPath) {
-        if (context.isTest()) {
-            logger.info("disable trace log during test");
-        } else {
-            context.logManager.traceLogger = TraceLogger.file(traceLogPath);
-        }
-    }
-
-    public void forwardLog(String kafkaURI) {
-        if (context.isTest()) {
-            logger.info("disable log forwarding during test");
-        } else {
-            context.logManager.logForwarder = LogForwarder.create(kafkaURI, context.logManager.appName);
-            context.stat.metrics.add(context.logManager.logForwarder.producerMetrics);
-            context.backgroundTask().scheduleWithFixedDelay(new CollectStatTask(context.logManager.logForwarder, context.stat), Duration.ofSeconds(10));
+            context.stat.metrics.add(context.logManager.kafkaAppender.producerMetrics);
+            context.backgroundTask().scheduleWithFixedDelay(new CollectStatTask(context.logManager.kafkaAppender, context.stat), Duration.ofSeconds(10));
         }
     }