+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Change log
 ### 6.0.0 (3/4/2018 - )             !!! only support Java 9+
 * jdk: drop java 8 support
+* log: removed write action/trace to file, updated sys.properties log key to sys.log.appender
+        sys.log.appender=console => write action/trace to console
+        sys.log.appender=kafkaURI => forward log to kafka
+       (in cloud env, console logging is prefered no matter it's docker or systemd/journald, or use log forwarding) 
 
 ### 5.3.8 (2/28/2018 - 3/4/2018)    !!! 5.3.X is last version to support Java 8
 * http: limit max requestURL length to 1000