+++ CHANGELOG.md
@@ -5,6 +5,7 @@
         sys.log.appender=console => write action/trace to console
         sys.log.appender=kafkaURI => forward log to kafka
        (in cloud env, console logging is prefered no matter it's docker or systemd/journald, or use log forwarding) 
+* http: update undertow to 2.0.1
 
 ### 5.3.8 (2/28/2018 - 3/4/2018)    !!! 5.3.X is last version to support Java 8
 * http: limit max requestURL length to 1000